pyproject.toml
[project]
name = "baike-correction"
version = "0.1.0"
description = "依照大百科校正文本"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.24.0",
    "numpy>=2.3.1",
    "openai>=1.93.1",
    "torch>=2.7.1",
    "transformers>=4.53.1",
]

[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
